# Skip SortExec for partitioning columns in OPTIMIZE